--STOCK WIZARD INSERT-- 1. Stok Ekleme Sihirbazı api ile çalışır hale getirildi 2. Listeye Filtreler eklendi 3. Bayi Detay içerisinde ki. kullanıcı listesinin API ile getirilmesi
--EXTRA--
1.CurrencyInput.js'de Separator Sembolü '.' olarak değiştirildi
2.ArrayColumnFormater.js eklendi -Sabitlerin Liste Ekranında Text olarak gösterilmesi-

+++ src/_metronic/layout/components/extras/PriceColumnFormatter.js
@@ -5,7 +5,7 @@ export const PriceColumnFormatter = (cellContent, row, rowIndex, { key }) => {
 
    return (
       <Currency
-         quantity={parseInt(row[key])}
+         quantity={parseFloat(row[key])}
          currency="TRY"
       />
    )
